Refactor Cryptography Manager
