+++ src/client/Microsoft.Identity.Client/Microsoft.Identity.Client.csproj
@@ -7,6 +7,19 @@
     <TargetFrameworkNetCore>netcoreapp2.1</TargetFrameworkNetCore>
     <TargetFrameworkNet5Win>net5.0-windows10.0.17763.0</TargetFrameworkNet5Win>
     
+    <TargetFrameworkNet6Android>net6.0-android</TargetFrameworkNet6Android>
+    <TargetFrameworkNet6Ios>net6.0-ios</TargetFrameworkNet6Ios>
+    <TargetFrameworkNet6Mac>net6.0-maccatalyst</TargetFrameworkNet6Mac>
+    <TargetFrameworkIos>Xamarin.iOS10</TargetFrameworkIos>
+    <TargetFrameworkMac>xamarinmac20</TargetFrameworkMac>
+    <TargetFrameworkAndroid9>MonoAndroid9.0</TargetFrameworkAndroid9>
+    <TargetFrameworkAndroid10>MonoAndroid10.0</TargetFrameworkAndroid10>
+
+    <TargetFrameworkNetStandard>netstandard1.3</TargetFrameworkNetStandard>
+    <!--<TargetFrameworkNetDesktop45>net45</TargetFrameworkNetDesktop45>-->
+    <!--<TargetFrameworkUap>uap10.0</TargetFrameworkUap>-->
+  
+
   </PropertyGroup>
 
   <!-- Mobile and legacy targets - comment these out or set env variable MSAL_DESKTOP_ONLY_DEV to speedup the build -->
@@ -175,8 +188,7 @@
     <!-- Manually include the cs files -->
     <Compile Include="$(PathToMsalSources)\**\*.cs" Exclude="$(PathToMsalSources)\obj\**\*.*" />
     <Compile Remove="$(PathToMsalSources)\Platforms\**\*.*;$(PathToMsalSources)\Resources\*.cs" />
-    <Compile Remove="$(PathToMsalSources)\PlatformsCommon\PlatformNotSupported\ApiConfig\SystemWebViewOptions.cs" />
-    <Compile Remove="GlobalSuppressions3.cs" />
+    <Compile Remove="$(PathToMsalSources)\PlatformsCommon\PlatformNotSupported\ApiConfig\SystemWebViewOptions.cs" />    
     <EmbeddedResource Include="$(PathToMsalSources)\Properties\Microsoft.Identity.Client.rd.xml" />
   </ItemGroup>
 
@@ -219,10 +231,6 @@
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == '$(TargetFrameworkNetDesktop461)' ">
-  </ItemGroup>
-
-  <ItemGroup Condition=" '$(TargetFramework)' == '$(TargetFrameworkNetDesktop45)' ">
-    <Compile Remove="$(PathToMsalSources)\PlatformsCommon\Shared\CryptographyManager.cs" />
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == '$(TargetFrameworkNetDesktop45)' Or '$(TargetFramework)' == '$(TargetFrameworkNetDesktop461)' ">
