--- src/client/Microsoft.Identity.Client/Platforms/Android/AndroidCryptographyManager.cs
@@ -1,57 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System;
-using System.Security.Cryptography;
-using System.Security.Cryptography.X509Certificates;
-using System.Text;
-using Microsoft.Identity.Client.Internal;
-using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
-using Microsoft.Identity.Client.Utils;
-
-namespace Microsoft.Identity.Client.Platforms.Android
-{
-#if MAUI
-    [Preserve(AllMembers = true)]
-#else
-    [global::Android.Runtime.Preserve(AllMembers = true)]
-#endif
-    internal class AndroidCryptographyManager : ICryptographyManager
-    {
-        public string CreateBase64UrlEncodedSha256Hash(string input)
-        {
-            return string.IsNullOrEmpty(input) ? null : Base64UrlHelpers.Encode(CreateSha256HashBytes(input));
-        }
-
-        public string GenerateCodeVerifier()
-        {
-            byte[] buffer = new byte[Constants.CodeVerifierByteSize];
-            using (RNGCryptoServiceProvider randomSource = new RNGCryptoServiceProvider())
-            {
-                randomSource.GetBytes(buffer);
-            }
-
-            return Base64UrlHelpers.Encode(buffer);
-        }
-
-        public string CreateSha256Hash(string input)
-        {
-            return string.IsNullOrEmpty(input) ? null : Convert.ToBase64String(CreateSha256HashBytes(input));
-        }
-
-        public byte[] CreateSha256HashBytes(string input)
-        {
-            using (var sha = new SHA256Managed())
-            {
-                return sha.ComputeHash(Encoding.UTF8.GetBytes(input));
-            }
-        }
-      
-        /// <inheritdoc />
-        public byte[] SignWithCertificate(string message, X509Certificate2 certificate)
-        {
-            // Used by Confidential Client, which is hidden on Android
-            throw new NotImplementedException();
-        }
-    }
-}
