+++ src/client/Microsoft.Identity.Client/Platforms/Mac/MacPlatformProxy.cs
@@ -106,7 +106,7 @@ namespace Microsoft.Identity.Client.Platforms.Mac
         }
 
         protected override IWebUIFactory CreateWebUiFactory() => new MacUIFactory();
-        protected override ICryptographyManager InternalGetCryptographyManager() => new MacCryptographyManager();
+        protected override ICryptographyManager InternalGetCryptographyManager() => new CommonCryptographyManager();
         protected override IPlatformLogger InternalGetPlatformLogger() => new ConsolePlatformLogger();
 
         protected override IFeatureFlags CreateFeatureFlags() => new MacFeatureFlags();
