+++ src/client/Microsoft.Identity.Client/Platforms/netcore/NetCoreDeviceAuthManager.cs
@@ -17,7 +17,7 @@ namespace Microsoft.Identity.Client.Platforms.netcore
 
         protected override byte[] SignWithCertificate(DeviceAuthJWTResponse responseJwt, X509Certificate2 certificate)
         {
-            return new NetCoreCryptographyManager().SignWithCertificate(responseJwt.GetResponseToSign(), certificate);
+            return new CommonCryptographyManager().SignWithCertificate(responseJwt.GetResponseToSign(), certificate);
         }
     }
 }
