+++ src/client/Microsoft.Identity.Client/Platforms/netdesktop/NetDesktopCryptographyManager.cs
@@ -19,41 +19,12 @@ using Microsoft.Win32.SafeHandles;
 
 namespace Microsoft.Identity.Client.Platforms.net45
 {
-    internal class NetDesktopCryptographyManager : ICryptographyManager
+    internal class NetDesktopCryptographyManager : CommonCryptographyManager
     {
         private static readonly ConcurrentDictionary<string, RSACryptoServiceProvider> s_certificateToRsaCspMap = new ConcurrentDictionary<string, RSACryptoServiceProvider>();
         private static readonly int s_maximumMapSize = 1000;
-
-        public string CreateBase64UrlEncodedSha256Hash(string input)
-        {
-            return string.IsNullOrEmpty(input) ? null : Base64UrlHelpers.Encode(CreateSha256HashBytes(input));
-        }
-
-        public string GenerateCodeVerifier()
-        {
-            byte[] buffer = new byte[Constants.CodeVerifierByteSize];
-            using (var randomSource = RandomNumberGenerator.Create())
-            {
-                randomSource.GetBytes(buffer);
-            }
-
-            return Base64UrlHelpers.Encode(buffer);
-        }
-
-        public string CreateSha256Hash(string input)
-        {
-            return string.IsNullOrEmpty(input) ? null : Convert.ToBase64String(CreateSha256HashBytes(input));
-        }
-
-        public byte[] CreateSha256HashBytes(string input)
-        {
-            using (var sha = SHA256.Create())
-            {
-                return sha.ComputeHash(Encoding.UTF8.GetBytes(input));
-            }
-        }
       
-        public byte[] SignWithCertificate(string message, X509Certificate2 certificate)
+        public override byte[] SignWithCertificate(string message, X509Certificate2 certificate)
         {
             if (certificate.PublicKey.Key.KeySize < CertificateClientCredential.MinKeySizeInBits)
             {
@@ -72,7 +43,7 @@ namespace Microsoft.Identity.Client.Platforms.net45
                 return signedData;
             }
 #else
-            return CryptographyManager.SignWithCertificate(message, certificate);
+            return base.SignWithCertificate(message, certificate);
 #endif
         }
