+++ src/client/Microsoft.Identity.Client/PlatformsCommon/Shared/CommonCryptographyManager.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using System.Collections.Concurrent;
 using System.Globalization;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
@@ -9,13 +10,20 @@ using System.Text;
 using Microsoft.Identity.Client.Internal;
 using Microsoft.Identity.Client.Internal.ClientCredential;
 using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
-using Microsoft.Identity.Client.PlatformsCommon.Shared;
 using Microsoft.Identity.Client.Utils;
 
-namespace Microsoft.Identity.Client.Platforms.netcore
+namespace Microsoft.Identity.Client.PlatformsCommon.Shared
 {
-    internal class NetCoreCryptographyManager : ICryptographyManager
+
+#if ANDROID && !MAUI
+    [global::Android.Runtime.Preserve(AllMembers = true)]
+#endif
+    [Preserve(AllMembers = true)]
+    internal class CommonCryptographyManager : ICryptographyManager
     {
+        private static readonly ConcurrentDictionary<string, RSA> s_certificateToRsaMap = new ConcurrentDictionary<string, RSA>();
+        private static readonly int s_maximumMapSize = 1000;
+
         public string CreateBase64UrlEncodedSha256Hash(string input)
         {
             return string.IsNullOrEmpty(input) ? null : Base64UrlHelpers.Encode(CreateSha256HashBytes(input));
@@ -44,18 +52,34 @@ namespace Microsoft.Identity.Client.Platforms.netcore
                 return sha.ComputeHash(Encoding.UTF8.GetBytes(input));
             }
         }
-   
+
         /// <inheritdoc />
-        public byte[] SignWithCertificate(string message, X509Certificate2 certificate)
+        public virtual byte[] SignWithCertificate(string message, X509Certificate2 certificate)
         {
+
+#if !NETSTANDARD            
             if (certificate.PublicKey.Key.KeySize < CertificateClientCredential.MinKeySizeInBits)
             {
                 throw new ArgumentOutOfRangeException(nameof(certificate),
                     string.Format(CultureInfo.InvariantCulture, MsalErrorMessage.CertificateKeySizeTooSmallTemplate,
                         CertificateClientCredential.MinKeySizeInBits));
             }
+#endif
 
-            return CryptographyManager.SignWithCertificate(message, certificate);
-        }
+            if (!s_certificateToRsaMap.TryGetValue(certificate.Thumbprint, out RSA rsa))
+            {
+                if (s_certificateToRsaMap.Count >= s_maximumMapSize)
+                    s_certificateToRsaMap.Clear();
+
+                rsa = certificate.GetRSAPrivateKey();
+            }
+
+            var signedData = rsa.SignData(Encoding.UTF8.GetBytes(message), HashAlgorithmName.SHA256, RSASignaturePadding.Pkcs1);
+
+            // Cache only valid RSA crypto providers, which are able to sign data successfully
+            s_certificateToRsaMap[certificate.Thumbprint] = rsa;
+            return signedData;
+
+        }        
     }
 }
